Move uri flags to util/flags/
